split rootReducer to multiple
