+++ src/services/actions/modalActions.js
@@ -1,0 +1,5 @@
+export const TOGGLE_MODAL = 'TOGGLE_MODAL';
+
+export const toggleModal = () => ({
+  type: TOGGLE_MODAL,
+});
